Find, Install, or Uninstall OpenMP

Set of functions that seek to identify whether OpenMP was installed, allow OpenMP to be installed, and removing OpenMP.

Usage

is_openmp_installed()

openmp_version()

openmp_install(
  password = base::getOption("macrtools.password"),
  verbose = TRUE,
  configure_makevars = TRUE
)

openmp_uninstall(
  password = base::getOption("macrtools.password"),
  verbose = TRUE,
  configure_makevars = TRUE
)

openmp_test()

Arguments

password

Password for user account to install software. Default is NULL.

verbose

Display messages indicating status

configure_makevars

Remove OpenMP configuration from ~/.R/Makevars during uninstall. Default is TRUE.

Details

The openmp_*() suite of functions attempts to locate, install, and uninstall OpenMP runtime library based on the Xcode version. OpenMP is installed to:

Library: /usr/local/lib/libomp.dylib

Headers: /usr/local/include/omp.h, /usr/local/include/ompt.h, /usr/local/include/omp-tools.h

Apple has explicitly disabled OpenMP support in Xcode compilers, but the runtime library can be installed separately and used with -Xclang -fopenmp compiler flags.

Check if OpenMP is installed

Checks the local file system for whether OpenMP runtime library is installed in the default installation location.

Installing OpenMP

The openmp_install() function aims to install the appropriate OpenMP runtime library based on the detected Xcode version.

OpenMP Installation Process

The installation process automatically detects your Xcode version and downloads the corresponding OpenMP runtime from the R-project repository:

# Download the appropriate version
curl -O https://mac.r-project.org/openmp/openmp-VERSION-darwin20-Release.tar.gz

# Install to system directories
sudo tar fvxz openmp-VERSION-darwin20-Release.tar.gz -C /

The installation includes:

/usr/local/lib/libomp.dylib - Main OpenMP runtime library

/usr/local/include/omp.h - OpenMP API header

/usr/local/include/ompt.h - OpenMP Tools interface header

/usr/local/include/omp-tools.h - OpenMP Tools header

Using OpenMP in R Packages

After installation, you can enable OpenMP support in R packages by adding the following to your ~/.R/Makevars file:

CPPFLAGS += -Xclang -fopenmp
LDFLAGS += -lomp

Or install packages with:

PKG_CPPFLAGS='-Xclang -fopenmp' PKG_LIBS=-lomp R CMD INSTALL myPackage

Uninstalling OpenMP

The openmp_uninstall() attempts to remove OpenMP from the default installation locations.

We use the R sanitized shell command to remove the installed files:

sudo rm -f /usr/local/lib/libomp.dylib
sudo rm -f /usr/local/include/omp.h
sudo rm -f /usr/local/include/ompt.h
sudo rm -f /usr/local/include/omp-tools.h
sudo rm -f /usr/local/include/ompx.h

Note that ompx.h was included in LLVM 19.1.0 and may not exist in older OpenMP versions.

Testing OpenMP Installation

After installing OpenMP, you can test if it's working by creating a simple test package or by adding the following to your ~/.R/Makevars file:

CPPFLAGS += -Xclang -fopenmp
LDFLAGS += -lomp

Then try installing a package that uses OpenMP, such as:

install.packages("data.table", type = "source")

Examples

# Check if OpenMP is installed
is_openmp_installed()
#> [1] FALSE