Find, Install, or Uninstall gfortran

Set of functions that seek to identify whether gfortran was installed, allow gfortran to be installed, and removing gfortran.

Usage

is_gfortran_installed()

gfortran_install(verbose = TRUE, password = NULL)

gfortran_uninstall(verbose = TRUE, password = NULL)

gfortran_update(verbose = TRUE, password = NULL)

Arguments

verbose

Display messages indicating status

password

Password for user account to install software. Default is NULL.

Check if gfortran is installed

Checks using the gfortran --version command to see if a response is present.

Uninstalling gfortran

The gfortran_uninstall() attempts to remove gfortran from two locations:

Intel (x86_64)

/usr/local/gfortran

/usr/local/bin/gfortran

M1/M2 (aarch64)

/opt/r/arm64/gfortran/

/opt/r/arm64/bin/gfortran

Using the R sanitized shell command of:

sudo -kS rm -r /usr/local/gfortran /usr/local/bin/gfortran

Or:

sudo -kS rm -r /opt/r/arm64/gfortran/ /opt/r/arm64/bin/gfortran

These uninstall steps are based on:

https://gcc.gnu.org/wiki/GFortranBinariesMacOS

And the desired path for aarch64:

https://mac.r-project.org/tools/

Examples

# Check if gfortran is installed
is_gfortran_installed()
#> [1] TRUE